separate support triggers from main package
