+++ tests/ahriman/application/handlers/test_handler_triggers_support.py
@@ -1,0 +1,1 @@
+# nothing to test here
